Move some new code into a new py_utils.cpp file
