+++ python/CMakeLists.txt
@@ -68,6 +68,7 @@ declare_mlir_python_extension(CUDAQuantumPythonSources.Extension
    runtime/cudaq/algorithms/py_optimizer.cpp
    runtime/cudaq/algorithms/py_sample_async.cpp
    runtime/cudaq/algorithms/py_state.cpp
+   runtime/cudaq/algorithms/py_utils.cpp
    runtime/cudaq/algorithms/py_vqe.cpp
    runtime/cudaq/platform/JITExecutionCache.cpp
    runtime/cudaq/platform/py_alt_launch_kernel.cpp
